Fix the image can not be load problem.
During local testing, i encountered an issue where the image did not initially render in p5.js.
This was due to the asynchronous nature of the loadImage() function.
To resolve this, we added a simple conditional check and console.log() messages within drawScreamImage() to verify the load status.
Once the image was confirmed to be loaded, it rendered correctly on the canvas.